fix(internal/surfer): remove request_id_field in non-Create commands (#3501)
Address discrepancies in the generated gcloud command surface by
refining the logic for mapping Protobuf fields to CLI arguments.

 ### Key Changes
- Fix `request_id_field` Logic: Conditioned the population of
request_id_field in the primary resource parameter to only apply for
Create methods. This prevents redundant or incorrect field mapping in
     Delete, Get, and List commands.
- Improved Type Mapping: Updated the generator to correctly map Protobuf
int64 and uint64 types to the long argparse type, ensuring compatibility
with gcloud framework conventions.

  Fixes: https://github.com/googleapis/librarian/issues/3500

---------

+++ internal/surfer/gcloud/utils/field_test.go
@@ -28,9 +28,9 @@ func TestGetGcloudType(t *testing.T) {
 	}{
 		{"String", api.STRING_TYPE, "str"},
 		{"Int32", api.INT32_TYPE, "int"},
-		{"Int64", api.INT64_TYPE, "int"},
+		{"Int64", api.INT64_TYPE, "long"},
 		{"UInt32", api.UINT32_TYPE, "int"},
-		{"UInt64", api.UINT64_TYPE, "int"},
+		{"UInt64", api.UINT64_TYPE, "long"},
 		{"Bool", api.BOOL_TYPE, "bool"},
 		{"Float", api.FLOAT_TYPE, "float"},
 		{"Double", api.DOUBLE_TYPE, "float"},
